+++ tango/integrations/torch/util.py
@@ -1,0 +1,48 @@
+import warnings
+from typing import TypeVar
+
+import torch
+from torch.utils.data import DistributedSampler, IterableDataset
+
+from .data import DataLoader
+
+T = TypeVar("T")
+
+
+def move_to_device(o: T, device: torch.device) -> T:
+    if isinstance(o, torch.Tensor):
+        return o.to(device)  # type: ignore[return-value]
+    elif isinstance(o, dict):
+        return {k: move_to_device(v, device) for k, v in o.items()}  # type: ignore[return-value]
+    elif isinstance(o, list):
+        return [move_to_device(x, device) for x in o]  # type: ignore[return-value]
+    elif isinstance(o, tuple):
+        return tuple((move_to_device(x, device) for x in o))  # type: ignore[return-value]
+    else:
+        return o
+
+
+def check_dataset(dataset, split: str):
+    try:
+        len(dataset)
+    except TypeError:
+        if not isinstance(dataset, IterableDataset):
+            warnings.warn(
+                f"Dataset for {split} split appears to be a streaming/iterable dataset, "
+                "but is not an instance of 'torch.utils.data.IterableDataset'. This could cause issues "
+                "within the DataLoader.",
+                UserWarning,
+            )
+
+
+def check_dataloader(dataloader: DataLoader):
+    # If using a regular dataset and not streaming/iterable dataset, we
+    # should probably be using a `DistributedSampler`.
+    if not isinstance(dataloader.dataset, IterableDataset) and not isinstance(
+        dataloader.sampler, DistributedSampler
+    ):
+        warnings.warn(
+            "DistributedSampler is required for dataloader during distributed training, "
+            f"found {type(dataloader.sampler)} instead.",
+            UserWarning,
+        )